262: Split ReactioTabs into separate component
