+++ ui/src/features/templates/TemplateHeader/TemplateHeader.tsx
@@ -28,7 +28,7 @@ import { addUpdateReactionField } from 'store/entities/reactions/reactions.thunk
 import { RemoveReaction } from 'features/reactions/RemoveReaction/RemoveReaction.tsx';
 import { selectTemplateById } from 'store/entities/templates/templates.selectors.ts';
 import { ReactionPreview } from 'features/reactions/ReactionPreview/ReactionPreview.tsx';
-import { downloadTemplateAsJson } from 'common/utils';
+import { downloadAsJson } from 'common/utils';
 
 interface TemplateHeaderProps {
   templateId: number;
@@ -58,8 +58,8 @@ export function TemplateHeader({ templateId, isReadyForEnumeration }: Readonly<T
     [templateId, location],
   );
 
-  const downloadAsJson = () => {
-    downloadTemplateAsJson(template, `${template.data.reactionId}.json`);
+  const downloadAsJsonHandle = () => {
+    downloadAsJson(template, `${template.data.reactionId}.json`);
   };
 
   return (
@@ -112,7 +112,7 @@ export function TemplateHeader({ templateId, isReadyForEnumeration }: Readonly<T
             <Button
               leftSection={<DownloadIcon />}
               variant="transparent"
-              onClick={downloadAsJson}
+              onClick={downloadAsJsonHandle}
             >
               Download Template in JSON
             </Button>
