+++ ui/src/pages/TemplatePage/TemplatePage.tsx
@@ -15,9 +15,9 @@
  */
 import { useParams } from 'wouter';
 import { useAppDispatch } from 'store/useAppDispatch.ts';
-import { type FC, Fragment, useEffect, useMemo } from 'react';
+import { useEffect, useMemo } from 'react';
 import { TemplateHeader } from 'features/templates/TemplateHeader/TemplateHeader.tsx';
-import { Badge, Flex, Paper, Tabs, Tooltip } from '@mantine/core';
+import { Badge, Flex, Paper } from '@mantine/core';
 import { useSelector } from 'react-redux';
 import classes from './TemplatePage.module.scss';
 import { ReactionDetailsSidebar } from 'features/reactions/ReactionDetailsSidebar/ReactionDetailsSidebar.tsx';
@@ -27,33 +27,7 @@ import { reactionEntityContext } from 'features/reactions/ReactionEntities/react
 import { CheckCircleIcon, CrossCircleIcon } from 'common/icons';
 import { getTemplate } from 'store/entities/templates/templates.thunks';
 import { selectTemplateById } from 'store/entities/templates/templates.selectors.ts';
-
-// test
-import type { ReactionViewSectionProps } from 'features/reactions/ReactionView/reactionView.types.ts';
-import { Inputs } from 'features/reactions/ReactionView/Inputs/Inputs.tsx';
-import { Identifiers } from 'features/reactions/ReactionView/Identifiers/Identifiers.tsx';
-// import { Notes } from 'features/reactions/ReactionView/Notes/Notes.tsx';
-import { Outcomes } from 'features/reactions/ReactionView/Outcomes/Outcomes.tsx';
-import { RequiredAsterisk } from 'common/components/display/RequiredAsterisk/RequiredAsterisk.tsx';
-
-interface ReactionTab {
-  name: string;
-  required?: true;
-  Component: FC<ReactionViewSectionProps>;
-}
-const createEmptyComponent = (name: string) => () => name;
-
-const tabs: Array<ReactionTab> = [
-  { name: 'inputs', required: true, Component: Inputs },
-  { name: 'outcomes', required: true, Component: Outcomes },
-  { name: 'conditions', Component: createEmptyComponent('conditions') },
-  { name: 'identifiers', Component: Identifiers },
-  { name: 'setup', Component: createEmptyComponent('setup') },
-  // { name: 'notes', Component: Notes },
-  { name: 'observations', Component: createEmptyComponent('observations') },
-  { name: 'workups', Component: createEmptyComponent('workups') },
-  { name: 'provenance', required: true, Component: createEmptyComponent('provenance') },
-];
+import { ReactionTabs } from 'features/reactions/ReactionEntities/ReactionTabs/ReactionTabs.tsx';
 
 export function TemplatePage() {
   const dispatch = useAppDispatch();
@@ -125,35 +99,7 @@ export function TemplatePage() {
               radius="md"
               p="lg"
             >
-              <Tabs
-                defaultValue={tabs[0].name}
-                classNames={{ tab: classes.tabTitle, panel: classes.panel }}
-              >
-                <Tabs.List>
-                  {tabs.map(({ name, required }) => (
-                    <Fragment key={name}>
-                      {required ? (
-                        <Tooltip label="Mandatory section">
-                          <Tabs.Tab value={name}>
-                            {name}
-                            <RequiredAsterisk />
-                          </Tabs.Tab>
-                        </Tooltip>
-                      ) : (
-                        <Tabs.Tab value={name}>{name}</Tabs.Tab>
-                      )}
-                    </Fragment>
-                  ))}
-                </Tabs.List>
-                {tabs.map(({ name, Component }) => (
-                  <Tabs.Panel
-                    key={name}
-                    value={name}
-                  >
-                    <Component reactionId={templateId} />
-                  </Tabs.Panel>
-                ))}
-              </Tabs>
+              <ReactionTabs reactionId={templateId} />
             </Paper>
             <ReactionDetailsSidebar reactionId={templateId} />
           </Flex>
