+++ ui/src/pages/TemplatePage/templatePage.module.scss
@@ -13,20 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-.tabTitle {
-  color: var(--color-text-secondary-3);
-  text-transform: uppercase;
-  font-weight: 500;
-
-  &[data-active] {
-    color: var(--color-text-primary);
-  }
-}
-
-.panel {
-  padding: 20px 20px 0;
-}
-
 .checkIcon {
   margin-right: 8px;
 }
